+++ src/test/java/com/yelp/nrtsearch/server/grpc/StateBackendServerTest.java
@@ -1652,7 +1652,7 @@ public class StateBackendServerTest {
             .setDefaultSearchTimeoutCheckEvery(500)
             .setDefaultTerminateAfter(5000)
             .setDefaultTerminateAfterMaxRecallCount(6000)
-            .setDeletePctAllowed(20)
+            .setDeletePctAllowed(20.0)
             .build();
 
     LiveSettingsResponse response = primaryClient.getBlockingStub().liveSettings(request);
@@ -1676,7 +1676,7 @@ public class StateBackendServerTest {
             .setVerboseMetrics(BoolValue.newBuilder().setValue(false).build())
             .setParallelFetchByField(BoolValue.newBuilder().setValue(false).build())
             .setParallelFetchChunkSize(Int32Value.newBuilder().setValue(50).build())
-            .setDeletePctAllowed(Int32Value.newBuilder().setValue(20).build())
+            .setDeletePctAllowed(DoubleValue.newBuilder().setValue(20.0).build())
             .build();
 
     IndexLiveSettings.Builder builder = IndexLiveSettings.newBuilder();
